+++ cpp/open3d/core/CMakeLists.txt
@@ -20,7 +20,7 @@ set (KERNEL_CUDA_SRC
 )
 
 set (CORE_SRC
-    DLPack/DLPackConverter.cpp
+    dlpack/DLPackConverter.cpp
     AdvancedIndexing.cpp
     ShapeUtil.cpp
     CUDAUtils.cpp
