+++ cpp/open3d/core/dlpack/DLPackConverter.cpp
@@ -24,11 +24,11 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/DLPack/DLPackConverter.h"
+#include "open3d/core/dlpack/DLPackConverter.h"
 
 #include "open3d/core/Blob.h"
-#include "open3d/core/DLPack/dlpack.h"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/dlpack/dlpack.h"
 
 namespace open3d {
 namespace core {
